api/requirements.txt
IPy 
PyYAML
flask
flask_cors
pandas
pandas_profiling
simplejson
aioflask
asgiref
pyyaml